PAY-35 add params to merchant model for calculation VAT and commissions

+++ database/migrations/20181127161408_commissions_collection.go
@@ -13,7 +13,7 @@ const (
 	pspCommission = 1.5
 )
 
-var pmCommissions = map[string]float32{
+var pmCommissions = map[string]float64{
 	"bank_card": 3,
 	"qiwi":      5,
 	"webmoney":  2,
